Cleanup separate concerns
