--- dom/cookie.js
@@ -1,3 +1,0 @@
-export default function cookie(name) {
-
-}
