--- dom/css.js
@@ -1,6 +1,0 @@
-
-export default function css (statics) {
-  return (...args) => {
-
-  }
-}
