--- dom/index.js
@@ -1,7 +1,0 @@
-import $ from './$.js'
-import attr from './attr.js'
-import input from './input.js'
-import html from './html.js'
-import h, { render } from './h.js'
-
-export { $, attr, input, html, render, h }
