+++ h.js
@@ -1,8 +1,8 @@
 import attr from './attr.js'
-import from, { observable, primitive } from '../core/from.js'
-import fx from '../core/fx.js'
-import prop from '../core/prop.js'
-import calc from '../core/calc.js'
+import from, { observable, primitive } from './from.js'
+import fx from './fx.js'
+import prop from './prop.js'
+import calc from './calc.js'
 
 const _group = Symbol('group'), _ptr = Symbol('ptr'), _props = Symbol('props'), _cleanup = Symbol('cleanup')
 const TEXT = 3, ELEMENT = 1
