+++ index.js
@@ -1,4 +1,9 @@
-import $ from './dom/$.js'
-export * from './dom/index.js'
-export * from './core/index.js'
+import $ from './$.js'
+import h, {render} from './h.js'
+import html from './html.js'
+import attr, {attrs} from './attr.js'
+import input from './input.js'
+
 export default $
+export {h, render, html, attr, attrs, input, $}
+export * from './core.js'
